% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\name{batlow_palettes}
\alias{batlow_palettes}
\alias{acton}
\alias{bam}
\alias{bam0}
\alias{bamako}
\alias{batlow}
\alias{batlowK}
\alias{batlowW}
\alias{berlin}
\alias{bilbao}
\alias{broc}
\alias{broc0}
\alias{buda}
\alias{bukavu}
\alias{cork}
\alias{corkO}
\alias{davos}
\alias{devon}
\alias{fes}
\alias{glasgow}
\alias{grayC}
\alias{hawaii}
\alias{imola}
\alias{lajolla}
\alias{lapaz}
\alias{lipari}
\alias{lisbon}
\alias{managua}
\alias{navia}
\alias{naviaW}
\alias{nuuk}
\alias{oleron}
\alias{oslo}
\alias{roma}
\alias{romaO}
\alias{tofino}
\alias{tokyo}
\alias{turku}
\alias{vanimo}
\alias{vik}
\alias{vikO}
\title{Batlow Palette Functions}
\usage{
acton(n)

bam(n)

bam0(n)

bamako(n)

batlow(n)

batlowK(n)

batlowW(n)

berlin(n)

bilbao(n)

broc(n)

broc0(n)

buda(n)

bukavu(n)

cork(n)

corkO(n)

davos(n)

devon(n)

fes(n)

glasgow(n)

grayC(n)

hawaii(n)

imola(n)

lajolla(n)

lapaz(n)

lipari(n)

lisbon(n)

managua(n)

navia(n)

naviaW(n)

nuuk(n)

oleron(n)

oslo(n)

roma(n)

romaO(n)

tofino(n)

tokyo(n)

turku(n)

vanimo(n)

vik(n)

vikO(n)
}
\arguments{
\item{n}{A single non-negative integer giving the number of colors to
generate.}
}
\value{
A character vector of hex colors.
}
\description{
Generate a vector of colors from one of the included batlow palettes.
}
\examples{
acton(8)
batlow(16)
vik(32)
}
