% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapColor.R
\name{mapColor}
\alias{mapColor}
\title{Map numeric values to colors}
\usage{
mapColor(x, xlim = NULL, pal = greyscale(16), clip = F)
}
\description{
Assign values (numeric) to a set of discrete colors linearly.
}
\author{
Thomas Bryce Kelly
}
\keyword{colormap}
\keyword{palette}
