% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{palette}
\alias{palette}
\title{palette
Color palette information.}
\format{
list of 40 named palettes containing a vector of 256 RGB color values.
}
\usage{
palette
}
\description{
palette
Color palette information.
}
\keyword{datasets}
